refactor: 폴더구조정리
