+++ src/features/auth/constants/input-fields.ts
@@ -1,0 +1,6 @@
+export const SIGN_UP_FIELDS = [
+  { id: 'username', label: '이름', type: 'text' },
+  { id: 'email', label: '이메일', type: 'email' },
+  { id: 'password', label: '비밀번호', type: 'password' },
+  { id: 'passwordConfirm', label: '비밀번호 확인', type: 'password' },
+] as const
