+++ src/features/pokemon/hooks/use-pokemon-modal.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import type { PokemonDetails } from '../types/pokemon'
+import type { PokemonDetails } from '../types/pokemon-api-types'
 
 export default function usePokemonModal() {
   const [selectedPokemon, setSelectedPokemon] = useState<PokemonDetails | null>(null)
