+++ src/features/pokemon/types/stats-types.ts
@@ -1,0 +1,12 @@
+export interface StatsTypes {
+  ID: number
+  name: string
+  img: string
+  flavorText: string
+  HP: number
+  attack: number
+  defense: number
+  specialAttack: number
+  specialDefense: number
+  speed: number
+}
